+++ Game/Include/Projectile.h
@@ -16,7 +16,7 @@ static const enum M_PROJECTILE_ITEM_SUB_TYPE
 class Projectile : public Item
 {
 public:
-	Projectile(const long id, const unsigned int itemType, const unsigned int itemSubType, float velocity, float damage);
+	Projectile(const unsigned int id, const unsigned int itemType, const unsigned int itemSubType, float velocity, float damage);
 	Projectile(const Projectile& other);
 	Projectile(const Projectile* other);
 	virtual ~Projectile();
