+++ Game/Source/RangedWeapon.cpp
@@ -1,6 +1,6 @@
 #include "RangedWeapon.h"
 
-RangedWeapon::RangedWeapon(const long ID, const unsigned int itemType, const int itemSubType, 
+RangedWeapon::RangedWeapon(const unsigned int ID, const unsigned int itemType, const int itemSubType, 
 						   const float damage, const float range) : Weapon(ID, itemType, itemSubType, damage, range)
 {
 	this->zStacking = false;
